rename save-assets.js to asset-saver.js
