+++ src/lib/asset-saver.js
